[IMP] website_event_sale: improve test naming and prepare for adding new tests
We rename the class according to the filename and extract the setup in a
specific class to allow to reuse the setup in other tests (see next commit).

Task-2802757
