+++ addons/website_event_sale/tests/__init__.py
@@ -3,4 +3,5 @@
 
 from . import common
 from . import test_frontend_buy_tickets
+from . import test_frontend_common
 from . import test_website_event_sale_pricelist
